requirements.txt
altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
eval_type_backport
gitdb
GitPython
h11
httpcore
httpx
idna
Jinja2
jsonschema
#jsonschema-specification
lxml
MarkupSafe
mistralai
narwhals
numpy
packaging
pandas
pillow
protobuf
pyarrow
pydantic
pydantic_core
pydeck
python-dateutil
python-docx
pytz
referencing
requests
rpds-py
six
smmap
sniffio
streamlit
tenacity
toml
tornado
typing-inspection
typing_extensions
tzdata
urllib3
watchdog
python-dotenv
html2docx
transformers
PyPDF2
FPDF
tabulate
plotly.express
pytest